Changed some names
I didn't change this to make it look like this is fully coded by me, but to look like this is my part of the project. If somebody wants to download it, then my project gets downloaded and not LWSS's Fuzion project. I have much respect for him maintaining the project, and I don't want to make his work look like mine. I'm sorry if that is the impression that I'm giving.